Refactor and split markup into components. Part 3
