+++ src/components/menu/components/section.js
@@ -1,8 +1,12 @@
-
-export const createMenuSection = (menu) => {
+/**
+ * Создание разметки меню
+ * @param {*} menuItems пункты меню
+ * @return {string} разметка меню
+ */
+export const createMenuSection = (menuItems) => {
   return (`
     <section class="control__btn-wrap">
-      ${menu}
+      ${menuItems}
     </section>
   `);
 };
