+++ src/components/tasks/task-edit-form/components/colorBarBlock.js
@@ -1,0 +1,9 @@
+export const createColorBar = () => {
+  return (`
+    <div class="card__color-bar">
+      <svg class="card__color-bar-wave" width="100%" height="10">
+        <use xlink:href="#wave"></use>
+      </svg>
+    </div>
+  `);
+};
