+++ src/components/tasks/task-edit-form/components/repeatingDaysBlock.js
@@ -1,0 +1,24 @@
+/**
+ * Создание шаблона формы выбора дня
+ * @param {*} days
+ * @param {*} repeatingDays
+ * @return {string}
+ */
+export const createRepeatingDaysMarkup = (days, repeatingDays) => {
+  return days.map((day, index) => {
+    const isChecked = repeatingDays[day];
+    return (`
+      <input
+        class="visually-hidden card__repeat-day-input"
+        type="checkbox"
+        id="repeat-${day}-${index}"
+        name="repeat"
+        value="${day}"
+        ${isChecked ? `checked` : ``}
+      />
+      <label class="card__repeat-day" for="repeat-${day}-${index}"
+        >${day}</label
+      >
+    `);
+  }).join(`\n`);
+};
